requirements.txt
telepot~=12.7
python-dotenv~=1.1.1
proxmoxer~=2.2.0